fix memory leak problem in dataloader

+++ tnt/utils/collate_fn.py
@@ -1,6 +1,7 @@
 import torch
 import numpy as np
 
+
 def multilabel_collate_fn(batch):
     batch.sort(key=lambda x: len(x[1]), reverse=True)
     img, label = zip(*batch)
@@ -11,5 +12,5 @@ def multilabel_collate_fn(batch):
         temp_label[:len(label[i])] = label[i]
         pad_label.append(torch.from_numpy(temp_label))
     img = torch.stack(img, 0)
-    pad_label = torch.stack(pad_label, 0)
+    pad_label = torch.stack(pad_label, 0).long()
     return img, pad_label
